Refactor: Main Mandalart View & Custom Drawer
