+++ lib/main.dart
@@ -20,10 +20,6 @@ class _MyAppState extends State<MyApp> {
   @override
   void initState() {
     super.initState();
-    _initControllers();
-  }
-
-  void _initControllers() {
     Get.put(DataController());
   }
