+++ pubspec.yaml
@@ -17,6 +17,7 @@ dependencies:
   sqflite: ^2.0.0+4
   path: ^1.8.0
   shared_preferences: ^2.0.12
+  fluttertoast: ^8.0.8
 
 dev_dependencies:
   flutter_test:
